tests/: revise
